requirements.txt
streamlit
opencv-python
numpy
Pillow
pandas
ultralytics
fast-plate-ocr
torch
torchvision
torchaudio